+++ src/combinatorics.h
@@ -1,6 +1,8 @@
 #ifndef _COMBINATORICS_H
 #define _COMBINATORICS_H 1
 
+#include <stdint.h>
+
 namespace math
 {
     
@@ -11,7 +13,7 @@ namespace math
  * @return silnia
  * @throws invalid_argument
  */
-unsigned long int factorial(const int n, const int k = 0);
+uint64_t factorial(const int n, const int k = 0);
 
 /**
  * Wariacja bez powtórzeń
@@ -19,12 +21,12 @@ unsigned long int factorial(const int n, const int k = 0);
  * @param k Wielkość kombinacji
  * @return Zwraca k-elementową wariację bez powtórzeń n-elemenowego zbioru
  */
-long double permutation(const int n, const int k);
+uint64_t permutation(const int n, const int k);
 
 /**
  * Kombinacja bez powtórzeń
  */
-long double combination(const int n, const int k);
+uint64_t combination(const int n, const int k);
 
 } /* namespace math */
